Conflicts:
	docs/sphinx/user-guide/release-notes/2.4.x.rst

+++ rel-eng/builder.py
@@ -21,22 +21,38 @@ opts = {
 mysession = koji.ClientSession("http://koji.katello.org/kojihub", opts)
 mysession.ssl_login(opts['cert'], opts['ca'], opts['serverca'])
 
-ARCH_LIST = ('i386', 'i686', 'x86_64')
-DISTRIBUTION_MAP = {
-    'el5': 'rhel5',
-    'el6': 'rhel6',
-    'fc19': 'fedora19',
-    'fc20': 'fedora20'
+ARCH = 'arch'
+REPO_NAME = 'repo_name'
+DIST_KOJI_NAME = 'koji_name'
+PULP_PACKAGES = 'pulp_packages'
+DISTRIBUTION_INFO = {
+    'el5': {
+        ARCH: ['i386', 'x86_64'],
+        REPO_NAME: '5Server',
+        DIST_KOJI_NAME: 'rhel5',
+        PULP_PACKAGES: ['pulp', 'pulp-rpm', 'pulp-puppet']
+    },
+    'el6': {
+        ARCH: ['i686', 'x86_64'],
+        REPO_NAME: '6Server',
+        DIST_KOJI_NAME: 'rhel6',
+        PULP_PACKAGES: ['pulp', 'pulp-nodes', 'pulp-rpm', 'pulp-puppet']
+    },
+    'fc19': {
+        ARCH: ['i686', 'x86_64'],
+        REPO_NAME: 'fedora-19',
+        DIST_KOJI_NAME: 'fedora19',
+        PULP_PACKAGES: ['pulp', 'pulp-nodes', 'pulp-rpm', 'pulp-puppet']
+    },
+    'fc20': {
+        ARCH: ['i686', 'x86_64'],
+        REPO_NAME: 'fedora-20',
+        DIST_KOJI_NAME: 'fedora20',
+        PULP_PACKAGES: ['pulp', 'pulp-nodes', 'pulp-rpm', 'pulp-puppet']
+    },
 }
 
-DISTRIBUTION_PUBLISH_DIRECTORIES = {
-    'el5': '5server',
-    'el6': '6server',
-    'fc19': 'fedora-19',
-    'fc20': 'fedora-20'
-}
-
-DIST_LIST = DISTRIBUTION_MAP.keys()
+DIST_LIST = DISTRIBUTION_INFO.keys()
 WORKSPACE = os.path.realpath(os.path.join(os.path.dirname(__file__), '../../'))
 MASH_DIR = os.path.join(WORKSPACE, 'mash')
 TITO_DIR = os.path.join(WORKSPACE, 'tito')
@@ -46,8 +62,14 @@ parser = argparse.ArgumentParser()
 parser.add_argument("version", help="The version of pulp to run the build for (2.4, 2.5, etc.)")
 parser.add_argument("stream", choices=['stable', 'testing', 'beta'],
                     help="The target release stream to build.")
-parser.add_argument("--packageonly", action="store_true", default=False,
-                    help="Create the repos only.  Do not build with koji.")
+parser.add_argument("--update-tag-package-list", action="store_true", default=False,
+                    help="Update the packages associated with the tag. This will verify that the "
+                         "dependencies & pulp packages are associated with the tag and that the "
+                         "specific versions of each of the dependencies has been associated with "
+                         "the appropriate tag. The current logged in user will be the owner for "
+                         "any packages that are added to the tag.")
+parser.add_argument("--disable-build", action="store_true", default=False,
+                    help="Disable koji building.")
 parser.add_argument("--disable-packaging", action="store_true", default=False,
                     help="Do not package or create repos")
 parser.add_argument("--push", action="store_true", default=False,
@@ -58,6 +80,8 @@ parser.add_argument("--build-dependency",
                     help="If specified, only build the specified dependency")
 parser.add_argument("--distribution", choices=DIST_LIST, nargs='+',
                     help="If specified, only build for the specified distribution")
+parser.add_argument("--tito-tag", help="The specific tito tag that should be built.  "
+                                       "If not specified the latest will be used.")
 
 opts = parser.parse_args()
 
@@ -85,6 +109,7 @@ if opts.build_dependency:
             DIST_LIST = dists_from_dep
     except IOError:
         print "dist_list.txt file not found for %s." % opts.build_dependency
+        sys.exit(1)
 
 
 # If a specific distribution or list of distributions has been specified on the command line
@@ -94,15 +119,18 @@ if opts.distribution:
     DIST_LIST = opts.distribution
 
 
-def ensure_dir(target_dir):
+def ensure_dir(target_dir, clean=True):
     """
-    Ensure that the directory specified exists and is empty.  This will delete the directory
-    if it already exists
+    Ensure that the directory specified exists and is empty.  By default this will delete
+    the directory if it already exists
 
     :param target_dir: The directory to process
     :type target_dir: str
+    :param clean: Whether or not the directory should be removed and recreated
+    :type clean: bool
     """
-    shutil.rmtree(target_dir, ignore_errors=True)
+    if clean:
+        shutil.rmtree(target_dir, ignore_errors=True)
     try:
         os.makedirs(target_dir)
     except OSError:
@@ -129,9 +157,13 @@ def build_srpm(distributions):
             distribution = ".%s" % dist
             print "Building Srpm for %s" % distribution
             command = ['tito', 'build', '--offline', '--srpm', '--output', tito_path,
-                                  '--dist', distribution]
+                       '--dist', distribution]
             if opts.scratch:
                 command.append('--test')
+
+            if opts.tito_tag:
+                command.append('--tag')
+                command.append(opts.tito_tag)
 
             subprocess.check_call(command)
 
@@ -154,7 +186,8 @@ def build_with_koji(build_tag_prefix, target_dists, scratch=False):
 
     for dist in target_dists:
         srpm_dir = os.path.join(TITO_DIR, dist)
-        build_target = "%s-%s" % (build_tag_prefix, DISTRIBUTION_MAP[dist])
+        build_target = "%s-%s" % (build_tag_prefix,
+                                  (DISTRIBUTION_INFO.get(dist)).get(DIST_KOJI_NAME))
 
         # Get all the source RPMs that were built
         # submit each srpm
@@ -216,31 +249,37 @@ def download_rpms_from_tag(tag, output_directory):
         location_on_koji = "%s%s" % (koji_url, koji_dir)
         command = "wget -r -np -nH --cut-dirs=4 -R index.htm*  %s -X %s" % \
                   (location_on_koji, data_dir)
-        print command
         subprocess.check_call(command, shell=True)
 
 
-def build_repos(output_dir):
+def build_repos(output_dir, dist):
     """
     Build the yum repos for each arch
     this method assumes that the directories already exist with each of the arches needed
 
     :param output_dir: The directory where the files were downloaded from koji
     :type output_dir: str
+    :param dist: The key for the distribution map for this repo
+    :type dist: str
     """
     noarch_dir = os.path.join(output_dir, 'noarch')
     comps_file = os.path.join(WORKSPACE, 'pulp', 'comps.xml')
-    for arch in ARCH_LIST:
+    arch_list = (DISTRIBUTION_INFO.get(dist)).get(ARCH)
+    for arch in arch_list:
         arch_dir = os.path.join(output_dir, arch)
-        if os.path.exists(arch_dir):
-            # Copy the noarch files to the arch_dir
-            command = "cp -R %s/* %s" % (noarch_dir, arch_dir)
-            print command
-            subprocess.check_call(command, shell=True)
-            # create the repo
-            command = "createrepo -g %s %s" % (comps_file, arch_dir)
-            subprocess.check_call(command, shell=True)
+        ensure_dir(arch_dir, False)
 
+        # Copy the noarch files to the arch_dir
+        command = "cp -R %s/* %s" % (noarch_dir, arch_dir)
+        subprocess.check_call(command, shell=True)
+        # create the repo
+        command = "createrepo -g %s %s" % (comps_file, arch_dir)
+        subprocess.check_call(command, shell=True)
+
+    # If there is an i686 directory rename it i386 since that is what the distributions expect
+    if os.path.exists(os.path.join(output_dir, 'i686')):
+        shutil.move(os.path.join(output_dir, 'i686'),
+                    os.path.join(output_dir, 'i386'))
     # Remove the noarch directory since it is not needed after repos are created
     shutil.rmtree(noarch_dir, ignore_errors=True)
 
@@ -280,6 +319,120 @@ def upload_and_unpack_binary_repository(target_directory, tar_file):
     command = 'rm %s/repo.tar' % (target_directory)
     run_destination_ssh_step(command)
 
+
+def get_tag_packages(tag):
+    """
+    Get the set of packages currently associated with a tag
+
+    :param tag: the tag to search for in koji
+    :type tag: str
+
+    :returns: a set of package names
+    :rtype: set of str
+    """
+    dsttag = mysession.getTag(tag)
+    pkglist = set([(p['package_name']) for p in mysession.listPackages(tagID=dsttag['id'])])
+    return pkglist
+
+
+def get_supported_dists_for_dep(dep_directory):
+    """
+    Get a list of the supported distributions for the dependency in the given directory
+
+    :param dep_directory: The full path of the directory where the dependency is stored
+    :type dep_directory: str
+
+    :returns: a set of dist keys for the dists that this dep supports
+    :rtype: set of str
+    """
+    dist_list_file = os.path.join(dep_directory, 'dist_list.txt')
+    try:
+        with open(dist_list_file, 'r') as handle:
+            line = handle.readline()
+            dists_from_dep = line.split(' ')
+    except IOError:
+        print "dist_list.txt file not found for %s." % dep_directory
+        sys.exit(1)
+
+    return set(dists_from_dep)
+
+
+def get_deps_for_dist(dist_key):
+    """
+    Get all the dependency packages that are required for a given distribution
+
+    :param dist_key: The distribution for which to get the deps
+    :type dist_key: str
+    :returns: a list of packages
+    :rtype: set of str
+    """
+    deps_packages = set([name for name in os.listdir(DEPS_DIR)
+                         if os.path.isdir(os.path.join(DEPS_DIR, name))])
+
+    dist_deps = []
+    for dep in deps_packages:
+        dist_list = get_supported_dists_for_dep(os.path.join(DEPS_DIR, dep))
+        if dist_key in dist_list:
+            dist_deps.append(dep)
+
+    return set(dist_deps)
+
+
+def add_packages_to_tag(base_tag):
+    """
+    Add a set of package to a given tag.  If a package needs to be added to the tag
+    the current user will be used as the owner.
+
+    :param base_tag: The distribution independent root of the build tag.  For example:
+                     pulp-2.4-testing
+    :type base_tag: str
+    """
+    task_list = []
+    current_user = mysession.getLoggedInUser().get('name')
+    # For each Distribution
+    for dist_key, dist_info in DISTRIBUTION_INFO.iteritems():
+        build_tag = "%s-%s" % (base_tag, dist_info.get(DIST_KOJI_NAME))
+        # get the current tagged builds for this build_tag
+        build_tag_existing_builds = [data.get('nvr') for data in mysession.listTagged(build_tag)]
+
+        dist_deps = get_deps_for_dist(dist_key)
+        packages_in_tag = get_tag_packages(build_tag)
+        delta = dist_deps.difference(packages_in_tag)
+        # add the regular dependencies
+        for package in delta:
+            # Add the package to the tag
+            print "adding %s to %s" % (package, build_tag)
+            mysession.packageListAdd(build_tag, package, current_user)
+
+        # verify all of the versions of the packages
+        for package in dist_deps:
+            # Get the current version of the dep and add it to the tag if it exists in koji
+            specfile = os.path.join(DEPS_DIR, package, "%s.spec" % package)
+            command = 'rpm --queryformat "%{RPMTAG_VERSION}-%{RPMTAG_RELEASE} "' \
+                      ' --specfile ' + specfile + ' | cut -f1 -d" "'
+            result = str(subprocess.check_output(command, shell=True))
+            version = result.strip()
+            # remove the distkey from the version string
+            version = version[:version.rfind('.')]
+            package_nvr = "%s-%s.%s" % (package, version, dist_key)
+            # Verify that this version of the package is not already in the tag
+            if package_nvr not in build_tag_existing_builds:
+                # verify that the package_nvr exists in koji at all
+                existing = mysession.search(package_nvr, 'build', 'glob')
+                if existing:
+                    print "Adding %s to %s" % (package_nvr, build_tag)
+                    task_id = mysession.tagBuild(build_tag, package_nvr)
+                    task_list.append(task_id)
+
+        # add the missing pulp packages
+        pulp_packages = set(dist_info.get(PULP_PACKAGES))
+        delta = pulp_packages.difference(packages_in_tag)
+        for package in delta:
+            print "adding %s to %s" % (package, build_tag)
+
+    # Monitor the task list until all are completed
+    wait_for_completion(task_list)
+
 # clean out and rebuild the mash directory & the target arch directories
 ensure_dir(MASH_DIR)
 ensure_dir(TITO_DIR)
@@ -290,8 +443,16 @@ os.makedirs(MASH_DIR)
 shutil.rmtree(TITO_DIR, ignore_errors=True)
 os.makedirs(TITO_DIR)
 
+# First update package tags if specified
+if opts.update_tag_package_list:
+    add_packages_to_tag(build_tag)
+    print "After updating packages the build target repositories need to rebuild.  Please monitor "\
+          "the koji server directly to ensure that those tasks have finished before building new " \
+          "dependencies or versions of pulp.  "
+    sys.exit(0)
+
 # Build the repos if requested
-if not opts.packageonly:
+if not opts.disable_build:
     build_srpm(DIST_LIST)
     builds = build_with_koji(build_tag, DIST_LIST, opts.scratch)
     wait_for_completion(builds)
@@ -299,15 +460,15 @@ if not opts.packageonly:
 # Don't build the repos if we are building a dependency
 if not opts.disable_packaging:
     # Download the rpms and create the yum repos
-    for distkey, distvalue in DISTRIBUTION_MAP.iteritems():
-        build_target = "%s-%s" % (build_tag, distvalue)
-        output_dir = os.path.join(MASH_DIR, DISTRIBUTION_PUBLISH_DIRECTORIES[distkey])
+    for distkey, distvalue in DISTRIBUTION_INFO.iteritems():
+        build_target = "%s-%s" % (build_tag, distvalue.get(DIST_KOJI_NAME))
+        output_dir = os.path.join(MASH_DIR, distvalue.get(REPO_NAME))
         ensure_dir(output_dir)
-        for arch in ARCH_LIST:
+        for arch in distvalue.get(ARCH):
             os.makedirs(os.path.join(output_dir, arch))
         print "Downloading tag: %s to %s" % (build_target, output_dir)
         download_rpms_from_tag(build_target, output_dir)
-        build_repos(output_dir)
+        build_repos(output_dir, distkey)
 
     # Push to hosted location
     if opts.push:
@@ -324,5 +485,3 @@ if not opts.disable_packaging:
         # run_destination_ssh_step(command)
         # upload_and_unpack_binary_repository(target_repo_dir, os.path.join(MASH_DIR, 'repo.tar'))
         # print "Finished updating repo"
-
-
